Param.hasChanged now always checked and set.
The value for Param.hasChanged was previously only set with a current value if the route was he same as the previous route (i.e., the same route has been reloaded with one or more params having been changed). Now the hasChanged property is always up-to-date.
The "before" and "leaving" function processing has also been modified: the use is still the same but if one handler function returns FALSE then the process ends immediately.